[REFACTOR] 입력 메서드 수정
