+++ src/main/java/view/InputView.java
@@ -24,15 +24,35 @@ public class InputView {
     public static int inputHeight() {
         System.out.println("최대 사다리 높이는 몇 개인가요?");
         int height = input.nextInt();
-        InputException.validateHeightSize(height);
+        validateHeightSize(height);
         input.nextLine();
         return height;
     }
 
+    public static void validateHeightSize(int height) {
+        if (height < 1) {
+            throw new IllegalArgumentException("사다리 높이는 1 이상이어야 합니다.");
+        }
+    }
+
     public static String inputViewerName() {
-        System.out.println("\n결과를 보고 싶은 사람은?");
-        String viewerName = input.nextLine();
-        InputException.validateViewerNameNotBlank(viewerName);
+        String viewerName;
+        while (true) {
+            System.out.println("\n결과를 보고 싶은 사람은?");
+            viewerName = input.nextLine();
+            try {
+                validateViewerNameNotBlank(viewerName);
+                break;
+            } catch (IllegalArgumentException e) {
+                System.out.println("입력 값이 공백일 수 없습니다. 다시 입력해 주세요.");
+            }
+        }
         return viewerName;
+    }
+
+    public static void validateViewerNameNotBlank(String viewerName) {
+        if (viewerName.isBlank()) {
+            throw new IllegalArgumentException("이름은 공백일 수 없습니다.");
+        }
     }
 }
